Modularize and Doxygen decklist.cpp
Took 1 hour 40 minutes

Took 11 seconds
