+++ cockatrice/src/client/tabs/tab_deck_storage.cpp
@@ -5,7 +5,7 @@
 #include "../../server/remote/remote_decklist_tree_widget.h"
 #include "../../settings/cache_settings.h"
 #include "../get_text_with_max.h"
-#include "decklist.h"
+#include "deck_list.h"
 #include "pb/command_deck_del.pb.h"
 #include "pb/command_deck_del_dir.pb.h"
 #include "pb/command_deck_download.pb.h"
