+++ cockatrice/src/client/ui/widgets/deck_analytics/mana_devotion_widget.h
@@ -6,7 +6,7 @@
 
 #include <QHBoxLayout>
 #include <QWidget>
-#include <decklist.h>
+#include <deck_list.h>
 #include <utility>
 
 class ManaDevotionWidget : public QWidget
