+++ cockatrice/src/deck/deck_stats_interface.cpp
@@ -1,6 +1,7 @@
 #include "deck_stats_interface.h"
 
-#include "decklist.h"
+#include "deck_list.h"
+#include "deck_list_card_node.h"
 
 #include <QDesktopServices>
 #include <QMessageBox>
