+++ tests/loading_from_clipboard/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_definitions("-DCARDDB_DATADIR=\"${CMAKE_CURRENT_SOURCE_DIR}/data/\"")
 add_executable(
-  loading_from_clipboard_test ../../common/decklist.cpp clipboard_testing.cpp loading_from_clipboard_test.cpp
+  loading_from_clipboard_test ../../common/deck_list.cpp clipboard_testing.cpp loading_from_clipboard_test.cpp
 )
 
 if(NOT GTEST_FOUND)
