+++ tests/loading_from_clipboard/clipboard_testing.h
@@ -1,7 +1,7 @@
 #ifndef CLIPBOARD_TESTING_H
 #define CLIPBOARD_TESTING_H
 
-#include "../../common/decklist.h"
+#include "../../common/deck_list.h"
 
 #include "gtest/gtest.h"
